Adding Routes to Feature Modules
In this commit, the app routing for the recipes paths is broken out into its own separate recipes-routing module. This module uses RouterModule.forChild as opposed to RouterModule.forRoot in the main routing module (one of these only). For child tells angular to append these routes to the root routes.